pyproject.toml
[project]
name = "openapi-viz"
version = "0.1.0"
description = "Visualization tool for OpenAPI schemas"
requires-python = ">=3.13"
dependencies = [
    "pyyaml",
    "graphviz",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
